move feedbacks loader from plugin to the itemRunner, register plugin and run it on the event
